More menu progress

+++ Assets/ScaleToMusic.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ScaleToMusic : MonoBehaviour {
 
-	public int bpm;
+	public float bpm;
 
 	// Use this for initialization
 	void Start () {
@@ -13,7 +13,7 @@ public class ScaleToMusic : MonoBehaviour {
 	
 	// Update is called once per frame
 	void Update () {
-		float scale = 1 + Mathf.Cos(Time.time * (bpm / 60) * Mathf.PI * 2);
+		float scale = 1f + Mathf.Cos(Time.time * (bpm / 60f) * Mathf.PI * 2f);
 		scale = Mathf.Lerp (0.95f, 1.05f, Mathf.InverseLerp (0f, 2f, scale));
 		transform.localScale = new Vector3(scale, scale, scale);
 	}
